testing/web-platform/mozilla/tests/indic-detection/shreetam.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
ö\ÆÁõ´ (Mars) `›¯USk®£zvÀ EÒÍ J¸ ÷PõÒ BS®. Cx `›¯Û¼¸¢x |õßPõÁx ÷PõÍõP EÒÍx.CUSk®£zvÀ ªPa]Ô¯ ÷PõÍõÚ ¦uİUS AkzuuõP CµshõÁx ]Ô¯ ÷PõÍõP ö\ÆÁõ´ C¸UQÓx. ÷©ÚõmiÚº CU÷PõĞUS ÷£õºUPhÄÎß ö£¯øµa `mi²ÒÍÚº. Cuß ÷©Ø£µ¨¤À Põn¨£k® C¸®¦ BUø\k CU÷PõøÍa ö\¢{Ó©õPU PõmkQÓx.[12] CuÚõ÷»÷¯ CuØSa ö\ÆÁõ´ GßÓ ö£¯º HØ£mhx. J¸ ¦Â\õº ÷PõÍõÚ Cuß ÷©Ø£µ¨¦ \¢vµÛÀ EÒÍx÷£õÀ QsnU SÈPøÍ²®, ¦Â°À EÒÍx ÷£õßÓ G›©ø»PÒ, £ÒÍzuõUSPÒ, £õø»ÁÚ[PÒ, £Û‰i¯ x¸Á¨ £SvPøÍ²® öPõshx. ö\ÆÁõ°ß _ÇØ]UPõ»•®, £¸Á ©õØÓ[PĞ® ¦ÂUS EÒÍøu¨ ÷£õßÓøÁ÷¯. `›¯ ©sh»zxÒ ªP E¯µ©õÚ J¼®£_ ©ø»²®, ªP¨ö£›¯ ö\[Szx¨ £ÒÍzuõUSPĞÒ JßÓõÚ ©›Úº £ÒÍzuõUS® ö\ÆÁõ°÷»÷¯ EÒÍÚ.
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://ta.wikipedia.org/w/index.php?title=%E0%AE%9A%E0%AF%86%E0%AE%B5%E0%AF%8D%E0%AE%B5%E0%AE%BE%E0%AE%AF%E0%AF%8D_(%E0%AE%95%E0%AF%8B%E0%AE%B3%E0%AF%8D)&oldid=3129711">&#2970;&#3014;&#2997;&#3021;&#2997;&#3006;&#2991;&#3021; (&#2965;&#3019;&#2995;&#3021;)</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
